Fixed a bug where average pace was not being properly converted from mins per km to mins per mile

+++ app/src/main/java/com/tjrushby/runlite/adapters/RunModelAdapter.java
@@ -50,7 +50,7 @@ public class RunModelAdapter extends RecyclerView.Adapter<RunModelViewHolder> {
         holder.setDateTime(formatter.dateToString(run.run.getDateTime()));
         holder.setTimeElapsed(formatter.intToHoursMinutesSeconds(run.run.getTimeElapsed()));
         holder.setDistance(formatter.doubleToDistanceStringWithUnits(run.run.getDistanceTravelled()));
-        holder.setAveragePace(formatter.doubleToAveragePaceString((long) run.run.getAveragePace()));
+        holder.setAveragePace(formatter.doubleToAveragePaceString(run.run.getAveragePace()));
 
         holder.itemView.setOnClickListener((View view) -> {
             if(view.getContext() instanceof MainActivity) {
